ref: MOve markup files into `trials/examples/trials`
