+++ src/JsPsych/trials/examples/partials/photodiode.js
