+++ src/JsPsych/trials/examples/photodiode/HoldUpMarker.js
@@ -1,7 +1,7 @@
 import htmlButtonResponse from '@jspsych/plugin-html-button-response';
 
-import { photodiodeGhostBox } from '../../../markup/photodiode';
-import { baseStimulus } from '../../../markup/stimuli';
+import { photodiodeGhostBox } from '../../../partials/photodiode';
+import { baseStimulus } from '../../../partials/stimuli';
 
 import { language } from '../../../language';
