+++ src/JsPsych/trials/examples/photodiode/StartCode.js
@@ -1,8 +1,8 @@
 import htmlKeyboardResponse from '@jspsych/plugin-html-keyboard-response';
 
 // TODO: Move markup to JsPsych?
-import { photodiodeGhostBox, pdSpotEncode } from '../../..//markup/photodiode';
-import { baseStimulus } from '../../..//markup/stimuli';
+import { photodiodeGhostBox, pdSpotEncode } from '../../../partials/photodiode';
+import { baseStimulus } from '../../../partials/stimuli';
 
 import { language } from '../../../language';
 import { EVENT_CODES, AUDIO_CODES } from '../../../constants';
